fix: use size_t to avoid narrowing conversion in get_size()

+++ network/packet/response/include/memory_data.hpp
@@ -9,7 +9,7 @@ namespace response
         memory_data(const std::vector<char> &data) : _data(data) {}
         std::vector<char> get_data() const { return _data; }
         std::vector<char> serialize() const;
-        int get_size() const;
+        size_t get_size() const;
 
     private:
         std::vector<char> _data;
