+++ network/packet/response/include/process_list_data.hpp
@@ -11,7 +11,7 @@ namespace response
         int get_process_count() const { return _process_count; }
         std::vector<process_info> get_processes() const { return _process_list; }
         std::vector<char> serialize() const;
-        int get_size() const;
+        size_t get_size() const;
 
     private:
         int _process_count;
