+++ network/packet/response/memory_data.cpp
@@ -5,7 +5,7 @@ std::vector<char> response::memory_data::serialize() const
     return _data;
 }
 
-int response::memory_data::get_size() const
+size_t response::memory_data::get_size() const
 {
     return _data.size();
 }
